+++ Port11.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Port11", "Port11\Port11.csp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitTests", "UnitTests\UnitTests.csproj", "{F5E63807-BB9A-429F-B28B-330FF65682AB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Settings", "Settings\Settings.csproj", "{FB25B370-43E7-4B04-9712-81D1354AD133}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -23,10 +21,6 @@ Global
 		{F5E63807-BB9A-429F-B28B-330FF65682AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F5E63807-BB9A-429F-B28B-330FF65682AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F5E63807-BB9A-429F-B28B-330FF65682AB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FB25B370-43E7-4B04-9712-81D1354AD133}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FB25B370-43E7-4B04-9712-81D1354AD133}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FB25B370-43E7-4B04-9712-81D1354AD133}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FB25B370-43E7-4B04-9712-81D1354AD133}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
