+++ Port11/Models/Header.cs
@@ -1,6 +1,6 @@
 ﻿namespace Port11.Models
 {
-    public class ResponseHeader
+    public class Header
     {
         public string Name { get; set; }
         public string Value { get; set; }
