+++ UnitTests/ModelsForTesting/Create.cs
@@ -1,0 +1,9 @@
+﻿namespace UnitTests.ModelsForTesting
+{
+    public class Create
+    {
+        public string name { get; set; }
+        public string salary { get; set; }
+        public string age { get; set; }
+    }
+}
